+++ package.json
@@ -26,6 +26,7 @@
   },
   "dependencies": {
     "npm": "^6.13.7",
+    "prop-types": "15.7.2",
     "react": "16.12.0",
     "react-dom": "16.12.0"
   }
